Fix accusations (#79)

+++ packages/hardhat-ts/contracts/BaseCase.sol
@@ -2,36 +2,36 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 contract BaseCase {
-  event CardDealt(uint card, address player);
-  event CardDiscarded(uint card, address player);
+  event CardDealt(uint256 card, address player);
+  event CardDiscarded(uint256 card, address player);
 
-  uint constant INVALID = 0;
+  uint256 constant INVALID = 0;
   // suspects
-  uint constant MUSTARD = 1;
-  uint constant SCARLET = 2;
-  uint constant PLUM = 3;
-  uint constant GREEN = 4;
-  uint constant WHITE = 5;
-  uint constant PEACOCK = 6;
+  uint256 constant MUSTARD = 1;
+  uint256 constant SCARLET = 2;
+  uint256 constant PLUM = 3;
+  uint256 constant GREEN = 4;
+  uint256 constant WHITE = 5;
+  uint256 constant PEACOCK = 6;
   // weapons
-  uint constant ROPE = 7;
-  uint constant PIPE = 8;
-  uint constant KNIFE = 9;
-  uint constant WRENCH = 10;
-  uint constant CANDLESTICK = 11;
-  uint constant REVOLVER = 12;
+  uint256 constant ROPE = 7;
+  uint256 constant PIPE = 8;
+  uint256 constant KNIFE = 9;
+  uint256 constant WRENCH = 10;
+  uint256 constant CANDLESTICK = 11;
+  uint256 constant REVOLVER = 12;
   // rooms
-  uint constant BILLIARD = 13;
-  uint constant STUDY = 14;
-  uint constant HALL = 15;
-  uint constant LOUNGE = 16;
-  uint constant DINING = 17;
-  uint constant BALLROOM = 18;
-  uint constant CONSERVATORY = 19;
-  uint constant LIBRARY = 20;
-  uint constant KITCHEN = 21;
+  uint256 constant BILLIARD = 13;
+  uint256 constant STUDY = 14;
+  uint256 constant HALL = 15;
+  uint256 constant LOUNGE = 16;
+  uint256 constant DINING = 17;
+  uint256 constant BALLROOM = 18;
+  uint256 constant CONSERVATORY = 19;
+  uint256 constant LIBRARY = 20;
+  uint256 constant KITCHEN = 21;
 
-  uint[] deck = [
+  uint256[] deck = [
     // suspects
     MUSTARD, SCARLET, PLUM, GREEN, WHITE, PEACOCK,
     // weapons
